# optimize addNodeToGroup by improving memory locality