Reduce the amount of memory accessed within `addNodeToGroup` by flattening the `nodes_c1c2s` array and storing the opposite node in the same array instead of looking it up in `state->transistors_c1` or `state->transistors_c2`.

In order to evaluate these changes, I added an `exit()` to the `CHRIN` routine to get a simple boot-to-BASIC benchmark.  Here are the results of timing this benchmark on my machine before these changes (running it three times to give an idea of the distribution):

real	0m2.815s
real	0m2.837s
real	0m2.798s

…and after these changes:

real	0m2.032s
real	0m1.987s
real	0m1.956s