@mist64 2020-06-04
Sorry for the delay, I missed the pull request. I can confirm the speedup:

Before:
2.25s user
2.14s user
2.20s user
2.23s user
2.19s user

After:
1.68s user
1.57s user
1.61s user
1.66s user
1.72s user

(It's enough to Ctrl+C once "READY." is reached; the program is blocked at that point.)

That's a 33% speedup!